requirements.txt
Django==5.1.1
djangorestframework>=3.14.0,<4.0
django-cors-headers>=3.14.0,<4.0
psycopg2-binary
djangorestframework-simplejwt

#react-router-dom
#npm install --save-dev @types/react
#npm install jwt-decode